drivers: wifi: nrf71: Clean up and relocate HAL layer
Remove legacy HAL event queues, tasklets, and IRQ workqueue now that
IPC delivers RPU events synchronously. Drop unused nRF70-era headers,
simplify nrf_wifi_hal_init(), and move HAL sources into the top-level
inc/ and src/ trees.
