+++ drivers/wifi/nrf71/Kconfig
@@ -255,21 +255,9 @@ config NRF71_TX_DONE_WQ_ENABLED
 config NRF71_RX_WQ_ENABLED
 	bool "RX workqueue"
 
-# Use for IRQ processing (TODO: using for BH processing causes issues)
-config NUM_METAIRQ_PRIORITIES
-	default 1
-
-config NRF71_IRQ_WQ_PRIORITY
-	int "Priority of the workqueue for handling IRQs"
-	default -15
-
 config NRF71_BH_WQ_PRIORITY
 	int "Priority of the workqueue for handling bottom half"
 	default 0
-
-config NRF71_IRQ_WQ_STACK_SIZE
-	int "Stack size of the workqueue for handling IRQs"
-	default 2048
 
 config NRF71_BH_WQ_STACK_SIZE
 	int "Stack size of the workqueue for handling bottom half"
