+++ drivers/wifi/nrf71/inc/offload_raw_tx/hal_api.h
@@ -13,7 +13,6 @@
 #ifndef __HAL_API_OFF_RAW_TX_H__
 #define __HAL_API_OFF_RAW_TX_H__
 
-#include <common/hal_structs_common.h>
 #include <common/hal_api_common.h>
 
 struct nrf_wifi_hal_dev_ctx *nrf_wifi_off_raw_tx_hal_dev_add(struct nrf_wifi_hal_priv *hpriv,
