+++ drivers/wifi/nrf71/osal/fw_if/umac_if/src/radio_test/fmac_api.c
@@ -227,7 +227,6 @@ void nrf_wifi_rt_fmac_dev_deinit(struct nrf_wifi_fmac_dev_ctx *fmac_dev_ctx)
 struct nrf_wifi_fmac_priv *nrf_wifi_rt_fmac_init(void)
 {
 	struct nrf_wifi_fmac_priv *fpriv = NULL;
-	struct nrf_wifi_hal_cfg_params hal_cfg_params;
 
 	fpriv = nrf_wifi_mem_zalloc(NRF_WIFI_MEM_POOL_TYPE_CTRL, sizeof(*fpriv));
 
@@ -237,13 +236,7 @@ struct nrf_wifi_fmac_priv *nrf_wifi_rt_fmac_init(void)
 		goto out;
 	}
 
-	nrf_wifi_mem_set(&hal_cfg_params,
-			      0,
-			      sizeof(hal_cfg_params));
-
-	fpriv->hpriv = nrf_wifi_hal_init(&hal_cfg_params,
-					 &nrf_wifi_rt_fmac_event_callback,
-					 NULL);
+	fpriv->hpriv = nrf_wifi_hal_init(&nrf_wifi_rt_fmac_event_callback);
 
 	if (!fpriv->hpriv) {
 		LOG_ERR("%s: Unable to do HAL init",
