--- drivers/wifi/nrf71/osal/hw_if/hal/inc/offload_raw_tx/phy_rf_params.h
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2025 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/**
- * @brief PHY init config parameters. These are passed to phy at init.
- */
-
-#ifndef _PHY_RF_PARAMS_H_
-#define _PHY_RF_PARAMS_H_
-#include "common/phy_rf_params_common.h"
-
-#define NRF_WIFI_OFF_RAW_TX_DEF_RF_PARAMS "007077003F032424001000002800323500000CF008087D8105010071630300EED501001F6F00003B350100F52E0000E35E0000B7B6000066EFFEFFB5F60000896200007A840200E28FFCFF080808080408120100000000A1A10178000000080050003B020726181818181A120A140E0600"
-
-#endif /* _PHY_RF_PARAMS_OFF_RAW_TX_H_ */
